Significant changes, specifically to the parent project that uses this library - inasmuch as this has now been broken out such that it can be maintained outside of that project.
Started writing test cases that are a mirror of the example cases of
IGES objects floating around on the web. It's been found that the
software that is used to test these IGES files does not support many
of the more complex features.

ToDo: Add a setup.py and add this to pip
      Doc strings
      Fix current object types in IGESGeomLib
      Continue implementing objects that are both useful and supported.